+++ eng/scripts/Language-Settings.ps1
@@ -12,14 +12,25 @@ function Get-java-PackageInfoFromRepo ($pkgPath, $serviceDirectory)
   {
     $projectData = New-Object -TypeName XML
     $projectData.load($projectPath)
+
+    if ($projectData.project.psobject.properties.name -notcontains "artifactId" -or !$projectData.project.artifactId) {
+      Write-Host "$projectPath doesn't have a defined artifactId so skipping this pom."
+      return $null
+    }
+
+    if ($projectData.project.psobject.properties.name -notcontains "version" -or !$projectData.project.version) {
+      Write-Host "$projectPath doesn't have a defined version so skipping this pom."
+      return $null
+    }
+
+    if ($projectData.project.psobject.properties.name -notcontains "groupid" -or !$projectData.project.groupId) {
+      Write-Host "$projectPath doesn't have a defined groupId so skipping this pom."
+      return $null
+    }
+
     $projectPkgName = $projectData.project.artifactId
     $pkgVersion = $projectData.project.version
-    if ($projectData.project.psobject.properties.name -contains "groupId") {
-      $pkgGroup = $projectData.project.groupId
-    }
-    else {
-      $pkgGroup = "unknown"
-    }
+    $pkgGroup = $projectData.project.groupId
 
     $pkgProp = [PackageProps]::new($projectPkgName, $pkgVersion.ToString(), $pkgPath, $serviceDirectory, $pkgGroup)
     if ($projectPkgName -match "mgmt" -or $projectPkgName -match "resourcemanager")