+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/AsyncDocumentClient.java
@@ -5,6 +5,9 @@ package com.azure.cosmos.implementation;
 import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.credential.TokenCredential;
 import com.azure.cosmos.ConsistencyLevel;
+import com.azure.cosmos.TransactionalBatchResponse;
+import com.azure.cosmos.implementation.apachecommons.lang.StringUtils;
+import com.azure.cosmos.implementation.batch.ServerBatchRequest;
 import com.azure.cosmos.CosmosPatchOperations;
 import com.azure.cosmos.TransactionalBatchResponse;
 import com.azure.cosmos.implementation.apachecommons.lang.StringUtils;
@@ -90,6 +93,7 @@ public interface AsyncDocumentClient {
         boolean sessionCapturingOverride;
         boolean transportClientSharing;
         boolean contentResponseOnWriteEnabled;
+        private CosmosClientMetadataCachesSnapshot state;
 
         public Builder withServiceEndpoint(String serviceEndpoint) {
             try {
@@ -97,6 +101,11 @@ public interface AsyncDocumentClient {
             } catch (URISyntaxException e) {
                 throw new IllegalArgumentException(e.getMessage());
             }
+            return this;
+        }
+
+        public Builder withState(CosmosClientMetadataCachesSnapshot state) {
+            this.state = state;
             return this;
         }
 
@@ -226,9 +235,10 @@ public interface AsyncDocumentClient {
                 tokenCredential,
                 sessionCapturingOverride,
                 transportClientSharing,
-                contentResponseOnWriteEnabled);
+                contentResponseOnWriteEnabled,
+                state);
 
-            client.init(null);
+            client.init(state, null);
             return client;
         }