+++ sdk/loganalytics/azure-resourcemanager-loganalytics/src/main/java/com/azure/resourcemanager/loganalytics/implementation/IntelligencePacksClientImpl.java
@@ -139,7 +139,6 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2020-08-01";
         return FluxUtil
             .withContext(
                 context ->
@@ -149,7 +148,7 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                             resourceGroupName,
                             workspaceName,
                             intelligencePackName,
-                            apiVersion,
+                            this.client.getApiVersion(),
                             this.client.getSubscriptionId(),
                             context))
             .contextWrite(context -> context.putAll(FluxUtil.toReactorContext(this.client.getContext()).readOnly()));
@@ -193,7 +192,6 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2020-08-01";
         context = this.client.mergeContext(context);
         return service
             .disable(
@@ -201,7 +199,7 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                 resourceGroupName,
                 workspaceName,
                 intelligencePackName,
-                apiVersion,
+                this.client.getApiVersion(),
                 this.client.getSubscriptionId(),
                 context);
     }
@@ -293,7 +291,6 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2020-08-01";
         return FluxUtil
             .withContext(
                 context ->
@@ -303,7 +300,7 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                             resourceGroupName,
                             workspaceName,
                             intelligencePackName,
-                            apiVersion,
+                            this.client.getApiVersion(),
                             this.client.getSubscriptionId(),
                             context))
             .contextWrite(context -> context.putAll(FluxUtil.toReactorContext(this.client.getContext()).readOnly()));
@@ -347,7 +344,6 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2020-08-01";
         context = this.client.mergeContext(context);
         return service
             .enable(
@@ -355,7 +351,7 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                 resourceGroupName,
                 workspaceName,
                 intelligencePackName,
-                apiVersion,
+                this.client.getApiVersion(),
                 this.client.getSubscriptionId(),
                 context);
     }
@@ -442,7 +438,6 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2020-08-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(
@@ -452,7 +447,7 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                             this.client.getEndpoint(),
                             resourceGroupName,
                             workspaceName,
-                            apiVersion,
+                            this.client.getApiVersion(),
                             this.client.getSubscriptionId(),
                             accept,
                             context))
@@ -492,7 +487,6 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2020-08-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service
@@ -500,7 +494,7 @@ public final class IntelligencePacksClientImpl implements IntelligencePacksClien
                 this.client.getEndpoint(),
                 resourceGroupName,
                 workspaceName,
-                apiVersion,
+                this.client.getApiVersion(),
                 this.client.getSubscriptionId(),
                 accept,
                 context);