+++ sdk/mediaservices/azure-resourcemanager-mediaservices/src/main/java/com/azure/resourcemanager/mediaservices/implementation/MediaServiceImpl.java
@@ -6,8 +6,8 @@ package com.azure.resourcemanager.mediaservices.implementation;
 
 import com.azure.core.http.rest.Response;
 import com.azure.core.management.Region;
+import com.azure.core.management.SystemData;
 import com.azure.core.util.Context;
-import com.azure.resourcemanager.mediaservices.MediaservicesManager;
 import com.azure.resourcemanager.mediaservices.fluent.models.MediaServiceInner;
 import com.azure.resourcemanager.mediaservices.models.AccountEncryption;
 import com.azure.resourcemanager.mediaservices.models.EdgePolicies;
@@ -17,7 +17,6 @@ import com.azure.resourcemanager.mediaservices.models.MediaServiceIdentity;
 import com.azure.resourcemanager.mediaservices.models.StorageAccount;
 import com.azure.resourcemanager.mediaservices.models.StorageAuthentication;
 import com.azure.resourcemanager.mediaservices.models.SyncStorageKeysInput;
-import com.azure.resourcemanager.mediaservices.models.SystemData;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -26,7 +25,7 @@ import java.util.UUID;
 public final class MediaServiceImpl implements MediaService, MediaService.Definition, MediaService.Update {
     private MediaServiceInner innerObject;
 
-    private final MediaservicesManager serviceManager;
+    private final com.azure.resourcemanager.mediaservices.MediaServicesManager serviceManager;
 
     public String id() {
         return this.innerModel().id();
@@ -94,7 +93,7 @@ public final class MediaServiceImpl implements MediaService, MediaService.Defini
         return this.innerObject;
     }
 
-    private MediaservicesManager manager() {
+    private com.azure.resourcemanager.mediaservices.MediaServicesManager manager() {
         return this.serviceManager;
     }
 
@@ -127,7 +126,7 @@ public final class MediaServiceImpl implements MediaService, MediaService.Defini
         return this;
     }
 
-    MediaServiceImpl(String name, MediaservicesManager serviceManager) {
+    MediaServiceImpl(String name, com.azure.resourcemanager.mediaservices.MediaServicesManager serviceManager) {
         this.innerObject = new MediaServiceInner();
         this.serviceManager = serviceManager;
         this.accountName = name;
@@ -157,7 +156,8 @@ public final class MediaServiceImpl implements MediaService, MediaService.Defini
         return this;
     }
 
-    MediaServiceImpl(MediaServiceInner innerObject, MediaservicesManager serviceManager) {
+    MediaServiceImpl(
+        MediaServiceInner innerObject, com.azure.resourcemanager.mediaservices.MediaServicesManager serviceManager) {
         this.innerObject = innerObject;
         this.serviceManager = serviceManager;
         this.resourceGroupName = Utils.getValueFromIdByName(innerObject.id(), "resourceGroups");