+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/main/java/com/azure/ai/metricsadvisor/implementation/models/PostgreSqlDataFeed.java
@@ -17,7 +17,7 @@ public final class PostgreSqlDataFeed extends DataFeedDetail {
     /*
      * The dataSourceParameter property.
      */
-    @JsonProperty(value = "dataSourceParameter", required = true)
+    @JsonProperty(value = "dataSourceParameter")
     private SqlSourceParameter dataSourceParameter;
 
     /**