+++ sdk/search/azure-search-documents/src/main/java/module-info.java
@@ -3,6 +3,7 @@
 
 module com.azure.search.documents {
     requires transitive com.azure.core;
+    requires transitive com.azure.core.experimental;
     requires jakarta.activation;
 
     opens com.azure.search.documents to com.fasterxml.jackson.databind;