+++ sdk/storage/azure-storage-blob-nio/src/test/java/com/azure/storage/blob/nio/CompositeTest.groovy
@@ -76,4 +76,21 @@ class CompositeTest extends APISpec {
         then:
         resultArr == defaultData.array()
     }
+
+    // Bug: https://github.com/Azure/azure-sdk-for-java/issues/20325
+    def "Files readAllBytes"() {
+        setup:
+        def fs = createFS(config)
+        def pathName = generateBlobName()
+        def path1 = fs.getPath("/foo/bar/" + pathName)
+        def path2 = fs.getPath("/foo/bar/" + pathName + ".backup")
+        Files.createFile(path1)
+        Files.createFile(path2)
+
+        when:
+        Files.readAllBytes(path1)
+
+        then:
+        notThrown(IOException)
+    }
 }