+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/storage/blob/ContainerAPITest.groovy
@@ -3,12 +3,15 @@
 
 package com.azure.storage.blob
 
+import com.azure.core.http.rest.PagedIterable
+import com.azure.core.http.rest.PagedResponse
 import com.azure.core.http.rest.Response
 import com.azure.identity.DefaultAzureCredentialBuilder
 import com.azure.storage.blob.models.AccessTier
 import com.azure.storage.blob.models.AppendBlobItem
 import com.azure.storage.blob.models.BlobAccessPolicy
 import com.azure.storage.blob.models.BlobErrorCode
+import com.azure.storage.blob.models.BlobItem
 import com.azure.storage.blob.models.BlobListDetails
 import com.azure.storage.blob.models.BlobProperties
 import com.azure.storage.blob.models.BlobRequestConditions
@@ -911,6 +914,27 @@ class ContainerAPITest extends APISpec {
             .verifyComplete()
     }
 
+    def "List blobs flat options maxResults by page"() {
+        setup:
+        def PAGE_SIZE = 2
+        def options = new ListBlobsOptions().setDetails(new BlobListDetails().setRetrieveCopy(true)
+            .setRetrieveSnapshots(true).setRetrieveUncommittedBlobs(true))
+        def normalName = "a" + generateBlobName()
+        def copyName = "c" + generateBlobName()
+        def metadataName = "m" + generateBlobName()
+        def tagsName = "t" + generateBlobName()
+        def uncommittedName = "u" + generateBlobName()
+        setupListBlobsTest(normalName, copyName, metadataName, tagsName, uncommittedName)
+
+        expect: "Get first page of blob listings (sync and async)"
+        for (def page : cc.listBlobs(options, null).iterableByPage(PAGE_SIZE)) {
+            assert page.value.size() <= PAGE_SIZE
+        }
+        StepVerifier.create(ccAsync.listBlobs(options).byPage(PAGE_SIZE).limitRequest(1))
+            .assertNext({ assert it.getValue().size() == PAGE_SIZE })
+            .verifyComplete()
+    }
+
     def "List blobs prefix with comma"() {
         setup:
         def prefix = generateBlobName() + ", " + generateBlobName()
@@ -1273,7 +1297,6 @@ class ContainerAPITest extends APISpec {
         !blobs.hasNext() // Normal
     }
 
-
     def "List blobs hier options maxResults"() {
         setup:
         def options = new ListBlobsOptions().setDetails(new BlobListDetails().setRetrieveCopy(true)
@@ -1289,6 +1312,26 @@ class ContainerAPITest extends APISpec {
         StepVerifier.create(ccAsync.listBlobsByHierarchy("", options).byPage().limitRequest(1))
             .assertNext({ assert it.getValue().size() == 1 })
             .verifyComplete()
+    }
+
+    def "List blobs hier options maxResults by page"() {
+        setup:
+        def options = new ListBlobsOptions().setDetails(new BlobListDetails().setRetrieveCopy(true)
+            .setRetrieveUncommittedBlobs(true))
+        def normalName = "a" + generateBlobName()
+        def copyName = "c" + generateBlobName()
+        def metadataName = "m" + generateBlobName()
+        def tagsName = "t" + generateBlobName()
+        def uncommittedName = "u" + generateBlobName()
+        setupListBlobsTest(normalName, copyName, metadataName, tagsName, uncommittedName)
+
+        expect:
+        def pagedIterable = cc.listBlobsByHierarchy("", options, null);
+
+        def iterableByPage = pagedIterable.iterableByPage(1)
+        for (def page : iterableByPage) {
+            assert page.value.size() == 1
+        }
     }
 
     @Unroll