+++ sdk/synapse/azure-analytics-synapse-artifacts/src/main/java/com/azure/analytics/synapse/artifacts/models/TumblingWindowFrequency.java
@@ -16,6 +16,9 @@ public final class TumblingWindowFrequency extends ExpandableStringEnum<Tumbling
     /** Static value Hour for TumblingWindowFrequency. */
     public static final TumblingWindowFrequency HOUR = fromString("Hour");
 
+    /** Static value Month for TumblingWindowFrequency. */
+    public static final TumblingWindowFrequency MONTH = fromString("Month");
+
     /**
      * Creates or finds a TumblingWindowFrequency from its string representation.
      *