+++ sdk/core/azure-core-management/CHANGELOG.md
@@ -3,6 +3,12 @@
 ## 1.3.0-beta.1 (Unreleased)
 
 
+## 1.2.1 (2021-04-02)
+
+### Dependency Updates
+
+- Upgraded `azure-core` from `1.14.0` to `1.15.0`.
+
 ## 1.2.0 (2021-03-08)
 
 ### New Features