+++ sdk/keyvault/azure-security-keyvault-certificates/CHANGELOG.md
@@ -1,7 +1,16 @@
 # Release History
 
-## 4.2.0-beta.5 (Unreleased)
+## 4.2.0-beta.6 (Unreleased)
 
+
+## 4.2.0-beta.5 (2021-04-09)
+
+### New features
+- Added support for service version `7.2`.
+- Added support to specify whether or not a pipeline policy should be added per call or per retry.
+
+### Breaking Changes
+- Changed `KeyVaultCertificateIdentifier` so it is instantiated via its constructor as opposed to via a `parse()` factory method.
 
 ## 4.2.0-beta.4 (2021-03-12)