+++ sdk/storage/azure-storage-queue/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 12.9.0-beta.3 (Unreleased)
+
+
 ## 12.9.0-beta.2 (2021-03-29)
 - Updated azure-storage-common and azure-core dependencies.